+++ plone/app/z3cform/tests/test_widgets.py
@@ -148,6 +148,7 @@ class DateWidgetTests(unittest.TestCase):
 
         self.request = TestRequest(environ={'HTTP_ACCEPT_LANGUAGE': 'en'})
         self.field = Date(__name__='datefield')
+        self.field.required = False
         self.widget = DateWidget(self.request)
         self.widget.field = self.field
         self.widget.pattern_options = {'date': {'firstDay': 0}}
@@ -164,7 +165,6 @@ class DateWidgetTests(unittest.TestCase):
                         'firstDay': 0,
                         'min': [current_year - 100, 1, 1],
                         'max': [current_year + 20, 1, 1],
-                        'clear': u'Clear',
                         'format': 'mmmm d, yyyy',
                         'monthsFull': [u'January', u'February', u'March',
                                        u'April', u'May', u'June', u'July',
@@ -175,18 +175,26 @@ class DateWidgetTests(unittest.TestCase):
                         'weekdaysFull': [u'Sunday', u'Monday', u'Tuesday',
                                          u'Wednesday', u'Thursday', u'Friday',
                                          u'Saturday'],
-                        'today': u'Today',
                         'selectYears': 200,
                         'placeholder': u'Enter date...',
                         'monthsShort': [u'Jan', u'Feb', u'Mar', u'Apr', u'May',
                                         u'Jun', u'Jul', u'Aug', u'Sep', u'Oct',
                                         u'Nov', u'Dec']
                     },
-                    'time': False
+                    'time': False,
+                    'today': u'Today',
+                    'clear': u'Clear',
                 }
             },
             self.widget._base_args(),
         )
+
+    def test_widget_required(self):
+        """Required fields should not have a "Clear" button.
+        """
+        self.field.required = True
+        base_args = self.widget._base_args()
+        self.assertEqual(base_args['pattern_options']['clear'], False)
 
     def test_data_converter(self):
         from plone.app.z3cform.widget import DateWidgetConverter
@@ -240,7 +248,9 @@ class DatetimeWidgetTests(unittest.TestCase):
 
         self.request = TestRequest(environ={'HTTP_ACCEPT_LANGUAGE': 'en'})
         self.field = Datetime(__name__='datetimefield')
+        self.field.required = False
         self.widget = DatetimeWidget(self.request)
+        self.widget.field = self.field
         self.widget.pattern_options = {
             'date': {'firstDay': 0},
             'time': {'interval': 15}
@@ -258,7 +268,6 @@ class DatetimeWidgetTests(unittest.TestCase):
                         'firstDay': 0,
                         'min': [current_year - 100, 1, 1],
                         'max': [current_year + 20, 1, 1],
-                        'clear': u'Clear',
                         'format': 'mmmm d, yyyy',
                         'monthsFull': [u'January', u'February', u'March',
                                        u'April', u'May', u'June', u'July',
@@ -269,7 +278,6 @@ class DatetimeWidgetTests(unittest.TestCase):
                         'weekdaysFull': [u'Sunday', u'Monday', u'Tuesday',
                                          u'Wednesday', u'Thursday', u'Friday',
                                          u'Saturday'],
-                        'today': u'Today',
                         'selectYears': 200,
                         'placeholder': u'Enter date...',
                         'monthsShort': [u'Jan', u'Feb', u'Mar', u'Apr', u'May',
@@ -278,14 +286,22 @@ class DatetimeWidgetTests(unittest.TestCase):
                     },
                     'time': {
                         'placeholder': u'Enter time...',
-                        'today': u'Today',
                         'format': 'h:i a',
                         'interval': 15
-                    }
+                    },
+                    'today': u'Today',
+                    'clear': u'Clear',
                 }
             },
             self.widget._base_args(),
         )
+
+    def test_widget_required(self):
+        """Required fields should not have a "Clear" button.
+        """
+        self.field.required = True
+        base_args = self.widget._base_args()
+        self.assertEqual(base_args['pattern_options']['clear'], False)
 
     def test_data_converter(self):
         from plone.app.z3cform.widget import DatetimeWidgetConverter